+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/Payload.java
@@ -91,7 +91,10 @@ public abstract class Payload {
     public enum PayloadType {
         TRANSFER,
         TRANSFER_WITH_MEMO,
-        REGISTER_DATA
+        REGISTER_DATA,
+        INIT_CONTRACT,
+        DEPLOY_MODULE,
+        UPDATE,
     }
 
 }