+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased changes
+- Added support for initializing smart contracts.
+- Added support for Update Contract.
+- Added support for Deploy Module transaction.
 
 - Support for P5
 - Fix a bug where pending changes for delegators were not visible via `getAccountInfo`.